crawler/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
psycopg[binary]==3.2.1
PyYAML==6.0.2
